4.1-translator_app/requirements.txt
langchain
langchain-openai
fastapi
uvicorn
langserve[all]
python-dotenv
